# code/openapi.yml
openapi: "3.1.0"
info:
  title: Climbing App API
  description: API for the Climbing Gym Boulder Creator app
  version: "1.0.0"
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/gyms:
    get:
      operationId: listGyms
      tags: [Gyms]
      summary: List all gyms
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated list of gyms
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GymPageResponse'
      security:
          - bearerAuth: []
    post:
      operationId: createGym
      tags: [Gyms]
      summary: Create a new gym (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGymRequest'
      responses:
        '201':
          description: Gym created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GymDto'
        '403':
          description: Forbidden - admin only
      security:
        - bearerAuth: []
  /api/gyms/{gymId}:
    get:
      operationId: getGymById
      tags: [Gyms]
      summary: Get gym by ID
      parameters:
        - name: gymId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Gym details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GymDto'
        '404':
          description: Gym not found
      security:
        - bearerAuth: []
    put:
      operationId: updateGym
      tags: [Gyms]
      summary: Update gym (admin or gym owner)
      parameters:
        - name: gymId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGymRequest'
      responses:
        '200':
          description: Gym updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GymDto'
        '403':
          description: Forbidden
        '404':
          description: Gym not found
      security:
        - bearerAuth: []
  /api/gyms/{gymId}/owners:
    post:
      operationId: assignGymOwner
      tags: [Gyms]
      summary: Assign a gym owner (admin only)
      parameters:
        - name: gymId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignOwnerRequest'
      responses:
        '201':
          description: Owner assigned
        '403':
          description: Forbidden - admin only
        '404':
          description: Gym or user not found
      security:
        - bearerAuth: []
  /api/gyms/{gymId}/wall-images:
    get:
      operationId: listWallImages
      tags: [Wall Images]
      summary: List wall images for a gym
      parameters:
        - name: gymId
          in: path
          required: true
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated list of wall image summaries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WallImagePageResponse'
      security:
        - bearerAuth: []
    post:
      operationId: uploadWallImage
      tags: [Wall Images]
      summary: Upload a wall image (admin or gym owner)
      parameters:
        - name: gymId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/WallImageUploadRequest'
      responses:
        '201':
          description: Image uploaded successfully
        '400':
          description: Invalid file
        '403':
          description: Forbidden
      security:
        - bearerAuth: []
  /api/wall-images/{wallImageId}:
    get:
      operationId: getWallImageById
      tags: [Wall Images]
      summary: Get wall image metadata
      parameters:
        - name: wallImageId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Wall image metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WallImageDetail'
        '404':
          description: Wall image not found
      security:
        - bearerAuth: []
  /api/wall-images/{wallImageId}/data:
    get:
      operationId: getWallImageData
      tags: [Wall Images]
      summary: Download full wall image
      parameters:
        - name: wallImageId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Wall image binary data
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
            image/webp:
              schema:
                type: string
                format: binary
        '404':
          description: Wall image not found
      security:
        - bearerAuth: []
  /api/wall-images/{wallImageId}/thumbnail:
    get:
      operationId: getWallImageThumbnail
      tags: [Wall Images]
      summary: Download wall image thumbnail
      parameters:
        - name: wallImageId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Thumbnail binary data
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
        '404':
          description: Wall image not found
      security:
        - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    GymDto:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        location:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        createdByAdminId:
          type: integer
        createdDate:
          type: string
          format: date-time
        updatedDate:
          type: string
          format: date-time
          nullable: true
    CreateGymRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        location:
          type: string
        description:
          type: string
        createdBy:
          type: integer
    UpdateGymRequest:
      type: object
      properties:
        name:
          type: string
        location:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
    AssignOwnerRequest:
      type: object
      required:
        - userId
        - assignedBy
      properties:
        userId:
          type: integer
        assignedBy:
          type: integer
    WallImageSummary:
      type: object
      properties:
        id:
          type: integer
        gymId:
          type: integer
        wallName:
          type: string
        isActual:
          type: boolean
        mimeType:
          type: string
        widthPx:
          type: integer
          nullable: true
        heightPx:
          type: integer
          nullable: true
        uploadedBy:
          type: integer
        uploadedDate:
          type: string
          format: date-time
        thumbnailUrl:
          type: string
    WallImageDetail:
      allOf:
        - $ref: '#/components/schemas/WallImageSummary'
        - type: object
          properties:
            imageUrl:
              type: string
    WallImageUploadRequest:
      type: object
      required:
        - image
        - wallName
        - uploadedBy
      properties:
        image:
          type: string
          format: binary
          description: Wall photo file (JPEG, PNG, or WebP)
        wallName:
          type: string
          description: Name of the physical wall (e.g. "Cave", "Slab Left")
        uploadedBy:
          type: integer
          description: ID of the user who uploaded the image
...
    PageResponse:
      type: object
      properties:
        content:
          type: array
          items: {}
        totalElements:
          type: integer
        totalPages:
          type: integer
        number:
          type: integer
        size:
          type: integer
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean
    GymPageResponse:
      allOf:
        - $ref: '#/components/schemas/PageResponse'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/GymDto'
    WallImagePageResponse:
      allOf:
        - $ref: '#/components/schemas/PageResponse'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/WallImageSummary'
...